% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.prop.func.R
\name{create.prop}
\alias{create.prop}
\title{Creates a list length J of category proportion for each cluster.}
\usage{
create.prop(
  J = J,
  q.vec = q.vec,
  Ktrue = Ktrue,
  strongprop = 0.8,
  which.noise = NULL
)
}
\arguments{
\item{J}{The number of active variable.!!!}

\item{q.vec}{A vector of length J giving the number of categories for each active variable.}

\item{Ktrue}{The number of clusters in J active variables.}

\item{strongprop}{A numeric value giving the strongest proportion of categories (common for all J active variables).}

\item{which.noise}{A vector of length (<= J) giving the index of noise variables in J active variables. NULL indicating all variable is non-noise.}
}
\value{
Returns a list length J, each of which is a (Ktrue x qj) matrix giving the proportion for each qj category in each Ktrue cluster.
}
\description{
Creates a list length J of category proportion for each cluster.
}
