% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.cate.func.R
\name{generate.cate.list}
\alias{generate.cate.list}
\title{Generate (NxJ) categorical data matrix.}
\usage{
generate.cate.list(N = N, prop.list = prop.list)
}
\arguments{
\item{N}{The number of observations.}

\item{prop.list}{a list length J, each of which is a vector of length qj giving the proportion for each categories.}
}
\value{
an (NxJ) categorical data matrix.
}
\description{
Generate an (NxJ) categorical data matrix given by prop.J.list and true cluster allocation.
}
