% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.ext.func.R
\name{generate.ext}
\alias{generate.ext}
\title{generates an artificial (NxH) external variable matrix.}
\usage{
generate.ext(N,extcate.vec=extcate.vec,unbala.cate=FALSE)
}
\arguments{
\item{N}{The number of observation.}

\item{extcate.vec}{A vector of length H, each element indicates the number of category for each H external variables.}

\item{unbala.cate}{logical value. If TRUE, the proportion of categories in the external variable is unbalanced. The default is FALSE.}
}
\value{
An (NxH) external variable matrix.
}
\description{
Generates an artificial (NxH) external variable matrix.
}
\examples{
###data setting
N <- 30 ; extcate.vec=c(2,3)
ext.mat=generate.ext(N,extcate.vec=extcate.vec)
}
\seealso{
\code{\link{generate.catecls}}
}
